requirements.txt
streamlit>=1.30.0
langchain>=0.2.2
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
groq>=0.5.0
PyMuPDF>=1.23.21  
pyodbc
python-dotenv
pyttsx3==2.90
qrcode==7.4.2
beautifulsoup4==4.12.3
pytesseract==0.3.10
langchain-experimental
langgraph
pypdf
